This seems closer to right, but socket is disconnecting
